Working Android version

+++ android/src/main/kotlin/com/bbflight/file_downloader/DownloadWorker.kt
@@ -1,7 +1,6 @@
 package com.bbflight.file_downloader
 
 import android.content.Context
-import android.content.SharedPreferences
 import android.os.Build
 import android.os.Handler
 import android.os.Looper
@@ -9,7 +8,6 @@ import android.util.Log
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
 import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
 import io.ktor.client.*
 import io.ktor.client.call.*
 import io.ktor.client.engine.cio.*
@@ -22,7 +20,6 @@ import kotlinx.coroutines.withContext
 import java.io.File
 import java.nio.file.Files
 import java.nio.file.StandardCopyOption
-import java.util.concurrent.locks.ReentrantLock
 import kotlin.io.path.Path
 import kotlin.io.path.pathString
 import kotlin.random.Random
@@ -37,7 +34,10 @@ class BackgroundDownloadTask(
     val baseDirectory: Int
 )
 
-/// Defines a set of possible states which a [DownloadTask] can be in.
+/** Defines a set of possible states which a [BackgroundDownloadTask] can be in.
+ *
+ * Must match the Dart equivalent enum, as value are passed as ordinal/index integer
+ */
 enum class DownloadTaskStatus {
     undefined,
     enqueued,
@@ -63,31 +63,7 @@ class DownloadWorker(
     companion object {
         const val TAG = "DownloadWorker"
         const val keyDownloadTask = "downloadTask"
-        const val keyTaskMap = "com.bbflight.file_downloader.taskMap"
-        const val keyNativeMap = "com.bbflight.file_downloader.nativeMap"
-        const val keyTaskIdMap = "com.bbflight.file_downloader.taskIdMap"
-
-        /** Get the native map from UserDefaults. Maps taskId to native id */
-        fun getNativeMap(prefs: SharedPreferences): MutableMap<String, String> {
-            val jsonString = prefs.getString(keyNativeMap, "{}")
-            val gson = Gson()
-            val mapType = object : TypeToken<MutableMap<String, String>>() {}.type
-            return gson.fromJson(jsonString, mapType)
-        }
-
-        /** Get the taskId map from TaskMap. Maps the native id to the taskId of the DownloadTask */
-        fun getTaskIdMap(prefs: SharedPreferences): MutableMap<String, String> {
-            val jsonString = prefs.getString(keyTaskIdMap, "{}")
-            val gson = Gson()
-            val mapType = object : TypeToken<MutableMap<String, String>>() {}.type
-            return gson.fromJson(jsonString, mapType)
-        }
     }
-
-    private val prefsSuccess = "com.bbflight.file_downloader.success"
-    private val prefsFailure = "com.bbflight.file_downloader.failure"
-    private val prefsLock = ReentrantLock()
-
 
     override suspend fun doWork(): Result {
         val gson = Gson()
@@ -106,7 +82,11 @@ class DownloadWorker(
     /** download a file from the urlString to the filePath */
     private suspend fun downloadFile(urlString: String, filePath: String): DownloadTaskStatus {
         try {
-            val client = HttpClient(CIO)
+            val client = HttpClient(CIO) {
+                install(HttpTimeout) {
+                    requestTimeoutMillis = 60000
+                }
+            }
             return client.prepareGet(urlString).execute { httpResponse ->
                 if (httpResponse.status.value in 200..299) {
                     withContext(Dispatchers.IO) {
@@ -117,29 +97,37 @@ class DownloadWorker(
                             dir
                         )
                         val channel: ByteReadChannel = httpResponse.body()
-                        while (!channel.isClosedForRead) {
+                        while (!channel.isClosedForRead && !isStopped) {
                             val packet = channel.readRemaining(DEFAULT_BUFFER_SIZE.toLong())
                             while (!packet.isEmpty) {
                                 val bytes = packet.readBytes()
                                 tempFile.appendBytes(bytes)
                             }
                         }
-                        val destFile = File(filePath)
-                        dir = destFile.parentFile
-                        if (!dir.exists()) {
-                            dir.mkdirs()
-                        }
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                            Files.move(
-                                tempFile.toPath(),
-                                destFile.toPath(),
-                                StandardCopyOption.REPLACE_EXISTING
-                            )
-                        } else {
-                            tempFile.copyTo(destFile, overwrite = true)
-                            tempFile.delete()
+
+                        if (!isStopped) {
+                            val destFile = File(filePath)
+                            dir = destFile.parentFile
+                            if (!dir.exists()) {
+                                dir.mkdirs()
+                            }
+                            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                                Files.move(
+                                    tempFile.toPath(),
+                                    destFile.toPath(),
+                                    StandardCopyOption.REPLACE_EXISTING
+                                )
+                            } else {
+                                tempFile.copyTo(destFile, overwrite = true)
+                                tempFile.delete()
+                            }
                         }
                     }
+                    if (isStopped) {
+                        Log.d(TAG, "Canceled task for $filePath")
+                        return@execute DownloadTaskStatus.canceled
+                    }
+                    Log.d(TAG, "Successfully downloaded to $filePath")
                     return@execute DownloadTaskStatus.complete
                 } else {
                     Log.w(
@@ -165,7 +153,7 @@ class DownloadWorker(
                 )
                 is CancellationException -> {
                     Log.i(TAG, "Job cancelled: $urlString to $filePath: ${e.message}")
-                    return DownloadTaskStatus.canceled;
+                    return DownloadTaskStatus.canceled
                 }
                 else -> Log.w(TAG, "Error downloading from $urlString to $filePath: ${e.message}")
             }
@@ -188,13 +176,13 @@ class DownloadWorker(
 
     /** Returns full path (String) to the file to be downloaded */
     private fun pathToFileForTask(task: BackgroundDownloadTask): String {
-        val baseDir = when (task.baseDirectory) {
-            0 -> applicationContext.dataDir
-            1 -> applicationContext.cacheDir
-            2 -> applicationContext.filesDir
+        val baseDirPath = when (task.baseDirectory) {
+            0 -> Path(applicationContext.dataDir.path, "app_flutter").pathString
+            1 -> applicationContext.cacheDir.path
+            2 -> applicationContext.filesDir.path
             else -> throw IllegalArgumentException("BaseDirectory int value ${task.baseDirectory} out of range")
         }
-        val path = Path(baseDir.path, task.directory)
+        val path = Path(baseDirPath, task.directory)
         return Path(path.pathString, task.filename).pathString
     }
